OAK-10141: refactored dynamic boost tests to be shared between lucene and elastic
